moved print headless to utils and removed formatted printing for safir-mem
