+++ safir-core/src/mem.rs
@@ -1,10 +1,9 @@
 use anyhow::Result;
 
 use async_trait::async_trait;
-use colored::*;
 
 use crate::config::SafirConfig;
-use crate::utils::{self, confirm_entry, print_header, print_output};
+use crate::utils::{self, confirm_entry, print_headless};
 use crate::SafirEngine;
 use rubin::net::client::RubinClient;
 
@@ -58,14 +57,13 @@ impl SafirEngine for SafirMemcache {
             return Ok(());
         }
 
-        print_header();
-        let output = if let Ok(val) = self.client.get_string(&key).await {
-            format!("{}: \"{}\"", key.bold().yellow(), val)
+        let value = if let Ok(val) = self.client.get_string(&key).await {
+            val
         } else {
-            format!("{}: ", key.bold().yellow())
+            String::from("")
         };
 
-        print_output(&output);
+        print_headless("", &key, &value);
 
         Ok(())
     }
@@ -89,16 +87,9 @@ impl SafirEngine for SafirMemcache {
             return;
         }
 
-        print_header();
-        let prefix = match prefix {
-            "alias" => "alias".bold().green(),
-            "export" => "export".bold().magenta(),
-            _ => prefix.bold(),
-        };
-
         for key in keys {
             if let Ok(value) = self.client.get_string(key).await {
-                println!("{} {}=\"{}\"\n", prefix, key.bold().yellow(), value);
+                print_headless(prefix, key, &value);
             }
         }
     }
